+++ firmware/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Show a more detailed error message on Vereinsflieger API errors
+
 ## 2.0.0 - 2026-03-16
 
 **Breaking Change**: Partition table has been updated to a new layout. Especially the config